requirements.txt
#numpy==1.26.4
pandas==2.1.4
python-dotenv==0.21.0
#mysql-connector-python==9.0.0
pymysql==1.1.1
scikit-learn==1.5.1
#seaborn==0.12.2
#xgboost==2.1.1
catboost==1.2.7
#mlflow==2.16.2
Flask==2.2.5 
#-e .
